requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cloudinary==1.38.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
mysqlclient==2.2.1
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
